Переносит логику getOffers в контроллер TripPoint
